// frontend/apps/web/src/domains/api/generated/model/jobRequirementRead.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Applark
 * OpenAPI spec version: 0.1.0
 */
import type { RequirementCategory } from './requirementCategory';
import type { JobRequirementReadEmbeddingModel } from './jobRequirementReadEmbeddingModel';

export interface JobRequirementRead {
  id: string;
  text: string;
  category: RequirementCategory;
  embedding_model: JobRequirementReadEmbeddingModel;
  created_at: string;
}