// frontend/apps/web/src/domains/api/generated/model/requirementCategory.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Applark
 * OpenAPI spec version: 0.1.0
 */

export type RequirementCategory = typeof RequirementCategory[keyof typeof RequirementCategory];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const RequirementCategory = {
  required: 'required',
  nice_to_have: 'nice_to_have',
  responsibility: 'responsibility',
} as const;